pyproject.toml
[tool.poetry]
name = "diffusion"
version = "0.1.0"
description = ""
authors = ["lee.chen <lee.chen@deltaww.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
torch = {version = "^2.5.1+cu121", source = "pytorch"}
torchvision = {version = "^0.20.1+cu121", source = "pytorch"}
torchaudio = {version = "^2.5.1+cu121", source = "pytorch"}
numpy = "1.24.4"
pillow = "^11.3.0"
segment-anything = {git = "https://github.com/facebookresearch/segment-anything.git"}
pyyaml = "^6.0.2"
opencv-contrib-python = "4.5.5.64"
requests = "^2.32.4"
packaging = "^25.0"
pycocotools = "^2.0.10"
timm = "^1.0.19"
transformers = "^4.54.0"

[[tool.poetry.source]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu121"
priority = "explicit"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
